With this feature branch the programme can now perform calculations on
reverse polish notaion expressions in the background, this feature has been implemented using TDD.

+++ calculator/src/main/java/calculator/NumStack.java
@@ -24,7 +24,7 @@ public class NumStack {
    * @param num The value to be converted to an {@link calculator.Entry Entry} object.
    */
   public void push(float num) {
-    Entry newEntry = new Entry((int) num);
+    Entry newEntry = new Entry(num);
     stack.push(newEntry);
   }
